tests: process58 fix killing of not running processes

+++ tests/probes/process58/command_line.sh
@@ -4,9 +4,9 @@ set -e -o pipefail
 
 function clean_processes {
 	# Processes are in stopped state. SIGCONT cause their exiting
-	kill -SIGCONT ${ZOMBIE_PPID}
-	kill -SIGCONT ${PID}
-	kill -SIGCONT ${ESCAPED_PID}  
+	[ -n "${ZOMBIE_PPID}" ] && kill -SIGCONT ${ZOMBIE_PPID}
+	[ -n "${PID}" ] && kill -SIGCONT ${PID}
+	[ -n "${ESCAPED_PPID}" ] && kill -SIGCONT ${ESCAPED_PID}
 }
 trap clean_processes EXIT
